+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2024-07-11
+
+### Added
+
+- upload() method for uploading firmware file
+- status messages will be sent to log
+
+### Changed
+
+- moved 'port' kwarg ahead of 'serial'
+- raise ChildProcessError instead of RuntimeError
+
 ## [0.1.2] - 2024-07-11
 
 ### Changed
@@ -23,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 _First release._
 
 
+[0.2.0]: https://github.com/int-brain-lab/tycmd-wrapper/releases/tag/v0.2.0
 [0.1.2]: https://github.com/int-brain-lab/tycmd-wrapper/releases/tag/v0.1.2
 [0.1.1]: https://github.com/int-brain-lab/tycmd-wrapper/releases/tag/v0.1.1
 [0.1.0]: https://github.com/int-brain-lab/tycmd-wrapper/releases/tag/v0.1.0